Separate tableView from ALKConversationViewController
