+++ Sources/Utilities/ALTopicDetail+Extension.swift
@@ -1,0 +1,41 @@
+//
+//  ALTopicDetail+Extension.swift
+//  ApplozicSwift
+//
+//  Created by Shivam Pokhriyal on 26/12/18.
+//
+
+import Foundation
+import Applozic
+
+extension ALTopicDetail: ALKContextTitleDataType {
+    public var titleText: String {
+        return title ?? ""
+    }
+
+    public var subtitleText: String {
+        return subtitle
+    }
+
+    public var imageURL: URL? {
+        guard let urlStr = link, let url = URL(string: urlStr) else {
+            return nil
+        }
+        return url
+    }
+
+    public var infoLabel1Text: String? {
+        guard let key = key1, let value = value1 else {
+            return nil
+        }
+        return "\(key): \(value)"
+    }
+
+    public var infoLabel2Text: String? {
+        guard let key = key2, let value = value2 else {
+            return nil
+        }
+        return "\(key): \(value)"
+    }
+
+}
